@@ -13,43 +13,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.2.2",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
     "@angular/pwa": "^0.13.0",
-    "@angular/router": "^7.2.2",
-    "@angular/service-worker": "^7.2.2",
+    "@angular/router": "^8.2.14",
+    "@angular/service-worker": "^8.2.14",
     "@ionic-native/core": "^5.0.0",
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
     "@ionic/angular": "4.0.1",
+    "@ng-select/ng-select": "^3.7.3",
+    "angularx-qrcode": "^2.1.0",
     "core-js": "^2.5.3",
     "firebase": "^5.8.2",
     "intersection-observer": "^0.5.0",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.29"
+    "rxjs": "~6.5.4",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "~7.2.2",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
     "@angular-devkit/architect": "~0.12.3",
-    "@angular-devkit/build-angular": "~0.12.3",
+    "@angular-devkit/build-angular": "~0.803.25",
     "@angular-devkit/core": "~7.2.3",
     "@angular-devkit/schematics": "~7.2.3",
-    "@ionic/ng-toolkit": "^1.0.0",
+    "@angular/cli": "~8.3.25",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
     "@ionic/angular-toolkit": "~1.2.0",
+    "@ionic/ng-toolkit": "^1.0.0",
     "@ionic/schematics-angular": "^1.0.0",
     "@types/node": "~10.9.2",
     "codelyzer": "~4.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "^3.4.5"
   },
   "description": "Build a Progressive Web App With Ionic 4"
 }
