@@ -1,23 +1,19 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
 
 const routes: Routes = [
-  {
-    path: '',
-    loadChildren: './home/home.module#HomePageModule'
-  },
-  {
-    path: 'settings',
-    loadChildren: './settings/settings.module#SettingsPageModule'
-  },
-  {
-    path: 'details',
-    loadChildren: './details/details.module#DetailsPageModule'
-  }
+    {
+        path: '',
+        loadChildren: './home/home.module#HomePageModule'
+    },
+    {
+        path: 'settings',
+        loadChildren: './settings/settings.module#SettingsPageModule'
+    }
 ];
 
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+@NgModule ({
+    imports: [RouterModule.forRoot (routes)],
+    exports: [RouterModule]
 })
 export class AppRoutingModule {}
