@@ -1,39 +1,39 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
 
-import { Platform } from '@ionic/angular';
-import { SplashScreen } from '@ionic-native/splash-screen/ngx';
-import { StatusBar } from '@ionic-native/status-bar/ngx';
+import {Platform} from '@ionic/angular';
+import {SplashScreen} from '@ionic-native/splash-screen/ngx';
+import {StatusBar} from '@ionic-native/status-bar/ngx';
 
-@Component({
-  selector: 'app-root',
-  templateUrl: 'app.component.html'
+@Component ({
+    selector: 'app-root',
+    templateUrl: 'app.component.html'
 })
 export class AppComponent {
-  public appPages = [
-    {
-      title: 'Home',
-      url: '/',
-      icon: 'home'
-    },
-    {
-      title: 'Settings',
-      url: '/settings',
-      icon: 'settings'
+    public appPages = [
+        {
+            title: 'Home',
+            url: '/',
+            icon: 'home'
+        },
+        {
+            title: 'Settings',
+            url: '/settings',
+            icon: 'settings'
+        }
+    ];
+
+    constructor (
+        private platform: Platform,
+        private splashScreen: SplashScreen,
+        private statusBar: StatusBar
+    ) {
+        this.initializeApp ();
     }
-  ];
 
-  constructor(
-    private platform: Platform,
-    private splashScreen: SplashScreen,
-    private statusBar: StatusBar
-  ) {
-    this.initializeApp();
-  }
-
-  initializeApp() {
-    this.platform.ready().then(() => {
-      this.statusBar.styleDefault();
-      this.splashScreen.hide();
-    });
-  }
+    initializeApp () {
+        this.platform.ready ().then (() => {
+            this.statusBar.styleDefault ();
+            this.splashScreen.hide ();
+        });
+    }
 }
