@@ -1,26 +1,27 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
-import { RouterModule } from '@angular/router';
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {RouterModule} from '@angular/router';
+import {IonicModule} from '@ionic/angular';
+import {NgSelectModule} from '@ng-select/ng-select';
+import {QRCodeModule} from 'angularx-qrcode';
 
-import { ComponentsModule } from '../components/components.module';
+import {HomePage} from './home.page';
 
-import { HomePage } from './home.page';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: HomePage
-      }
-    ]),
-    ComponentsModule
-  ],
-  declarations: [HomePage]
+@NgModule ({
+    imports: [
+        CommonModule,
+        FormsModule,
+        IonicModule,
+        NgSelectModule,
+        QRCodeModule,
+        RouterModule.forChild ([
+            {
+                path: '',
+                component: HomePage
+            }
+        ])
+    ],
+    declarations: [HomePage]
 })
 export class HomePageModule {}
